Fix : 테스트 코드users, auth 수정 완료
