+++ src/users/users.service.js
@@ -8,12 +8,12 @@ export class UsersService {
         const permisson = await this.usersRepository.perMisson(userId);
         const findAllUser = await this.usersRepository.findAllUser();
         
-        if(findAllUser.length === 0){
-            throw new Error("현재 존재하는 유저가 없습니다.");
-        }
-
         if(permisson.authorityName !== "ADMIN"){
             throw new Error("관리자 권한이 존재하지 않아 접근이 불가합니다.");
+        }
+
+        if(findAllUser.length === 0){
+            throw new Error("현재 존재하는 유저가 없습니다.");
         }
 
         return findAllUser;
